requirements.txt
# Hurricane Debris Detection - Dependencies
# Install: pip install -r requirements.txt

# Core
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0

# Vision & Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.1

# Hugging Face / Florence-2
transformers>=4.38.0
peft>=0.8.0

# SAM2
# Install separately: pip install git+https://github.com/facebookresearch/segment-anything-2.git

# Evaluation
pycocotools>=2.0.7
torchmetrics>=1.2.0

# Data & Utilities
tqdm>=4.65.0
requests>=2.31.0
gdown>=4.7.1
pyyaml>=6.0.1

# Logging & Experiment Tracking
tensorboard>=2.15.0

# Deployment
gradio>=4.44.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
